+++ tests/Unit/Contrib/Otlp/OtlpUtilTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace OpenTelemetry\Tests\Unit\Contrib\Otlp;
 
+use OpenTelemetry\API\Common\Signal\Signals;
 use OpenTelemetry\Contrib\Otlp\OtlpUtil;
 use PHPUnit\Framework\TestCase;
 
@@ -17,5 +18,29 @@ class OtlpUtilTest extends TestCase
         $header = OtlpUtil::getUserAgentHeader();
         $this->assertArrayHasKey('User-Agent', $header);
         $this->assertNotNull($header['User-Agent']);
+    }
+
+    public function test_method_not_defined(): void
+    {
+        $this->expectException(\UnexpectedValueException::class);
+        OtlpUtil::method('foo');
+    }
+
+    /**
+     * @dataProvider methodProvider
+     */
+    public function test_method(string $signal, string $expected): void
+    {
+        $method = OtlpUtil::method($signal);
+        $this->assertStringContainsString($expected, $method);
+    }
+
+    public function methodProvider(): array
+    {
+        return [
+            [Signals::TRACE, 'TraceService'],
+            [Signals::METRICS, 'MetricsService'],
+            [Signals::LOGS, 'LogsService'],
+        ];
     }
 }
